third_party/WebKit/LayoutTests/http/tests/loading/sync-xhr-priority-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
http://127.0.0.1:8000/js-test-resources/js-test-pre.js has priority Medium
http://127.0.0.1:8000/resources/test.mp4 has priority VeryHigh
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
PASS successfullyParsed is true

TEST COMPLETE
